Corregidos DTOs

+++ src/main/java/com/curso/dto/RentDto.java
@@ -1,7 +1,7 @@
 package com.curso.dto;
 
 public class RentDto {
-	public RentDto(Integer id, UserDto user, CarDto car, Integer initDate, Integer finalDate, Double price) {
+	public RentDto(Integer id, UserDto user, CarDto car, Long initDate, Long finalDate, Double price) {
 		this.id = id;
 		this.user = user;
 		this.car = car;
@@ -12,8 +12,8 @@ public class RentDto {
 	private Integer id;
 	private UserDto user;
 	private CarDto car;
-	private Integer initDate;
-	private Integer finalDate;
+	private Long initDate;
+	private Long finalDate;
 	private Double price;
 	public UserDto getUser() {
 		return user;
@@ -27,16 +27,16 @@ public class RentDto {
 	public void setCar(CarDto car) {
 		this.car = car;
 	}
-	public Integer getInitDate() {
+	public Long getInitDate() {
 		return initDate;
 	}
-	public void setInitDate(Integer initDate) {
+	public void setInitDate(Long initDate) {
 		this.initDate = initDate;
 	}
-	public Integer getFinalDate() {
+	public Long getFinalDate() {
 		return finalDate;
 	}
-	public void setFinalDate(Integer finalDate) {
+	public void setFinalDate(Long finalDate) {
 		this.finalDate = finalDate;
 	}
 	public Double getPrice() {
